Starting on Issue #7
Framing in terms of LoL stats rather than a random dataset. I'm not seeing a very strong relationship between the data, at least when it comes to armor and health,
but I want to experiment to what extent the anomalies are introduced, then experiment with the contamination rate of the Isolation Forest to see how accurate
we can differentiate between real data and anomalies.